requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
pytesseract>=0.3.8
Pillow>=8.0.0
python-dotenv>=0.15.0
